Grib reader (#72)
## Purpose

Refactor the grib reader so that parameters can be loaded in individual functions that can later be dask delayed. 


## Code changes:
- implements the grib reader with a class instead of standalone functions
- introduce the option to dask delay the load functions
- reference field is loaded first, before any other parameter. 
- one record found of the reference parameter is enough to derive the grid reference information, therefore the full field is not constructed. 
- implement a Grid class to store the grid reference attributes

## Performance impact
these changes introduce a performance deterioration. The jenkins tests increased timing from ~3.30m to ~6.30m. 
This is due to the fact that with earthkit-data it is more efficient to load a collection of data fields at once, i.e. traversing only once the sequence of grib records contained in a grib file. That means, random access to fields is less efficient. And this PR split the load of a data field collection into individual tasks each loading one field, requiring then random access to data field loads. 
However, the final architecture for performance should be measured against a system that can efficiently provide random access  to data fields, i.e. FDB. Additionally, we should measure the effect on an entire product collection parallelize. Since often when parallelizing an algorithm (to make it thread safe or parallel) it incurs into a performance hit on the sequential code performance. 
I quickly measured the impact using the indexing of earhkit-data (#75 & #76). But I observed no difference