First product (#73)
## Purpose

Introduce a base class for products. The purpose in this PR is to provide through that base class infrastructure to automatically dask delay product runs. In the future it can also be used for other common infrastructure for products like logging, or perf monitoring (measure run of product)

Products should then follow a contract with the base class, like they should specify input_fields attribute, that list the input fields that need to be loaded. Today that is a collection of names, but in the future the data requests specification will probably require more complex requests (like mars requests).

## Code changes:

- implementation of product base class
- load function in the grib_decoder that accepts a collection of products (from where the input data collection is extracted).