Metadata grid (#116)
## Purpose

Refactor logic that sets the horizontal components of the origin attribute into the metadata module. Loading the reference grid is no longer done during the initialisation of the GribReader instance but can be triggered by a dedicated function. This notably avoids hitting the data source (e.g. FDB) with a request of a single field.

## Code changes:

- Added `metadata.set_origin_xy` function
- Deprecated GribReader ref_param argument