Drive vcoord_type and origin_z from grib message (#120)
## Purpose

Use the `typeOfLevel` key as the source of truth for the `vcoord_type` attribute. An `extract` function is added to the metadata module to enable this workflow. The `override` function now updates all attributes previously extracted by the `grib_decoder` module. This does not include the `origin_x` and `origin_y` attributes as they need to be derived from a reference field.

## Code changes

- Added `metadata.extract`
- `metadata.override` will call the `extract` function to update attributes in the return value