Refactor grib reader (#122)
## Purpose

The GribReader only accepts requests for a single paramter at a time. The requests are then dispatched through the tasking mechanism. This is however unfortunate for the polytope datasource as every separate request is queued and processed separately. The present pull request implements a lower level interface that can deal with requests for multiple params. The reader interface remains unchanged.

## Code changes:

- Added `grib_decoder.load` and `grib_decoder.load_single_param` functions.