Add step unit and ref_time dim (#40)
## Purpose

The update to eccodes 2.35 introduced the support for step units in the mars language. This has a side effect that the step key may return a string of the value with a symbol for the unit of time. This PR introduces the `lead_time` dimension with coordinates encoded using timedelta objects in order to support the step units. Additionally, the `ref_time` is promoted to a dimension to accommodate use cases where multiple reference times are loaded from a data source.

## Code changes:

- dimension `time` renamed to `lead_time` and type changed from `int` to `timedelta64[ns]`
- added dimension `ref_time` with type `datetime64[ns]`