Set GRIB metadata for UTM grids. Otherwise set geography to "does not apply" (#66)
## Purpose

When regridding from ICON, add GRIB geography support for UTM grids in
northern hemisphere and add a default case that sets the grid definition
to "does not apply". The current behaviour leaves the definition the
same, which puts the xarray in an inconsistent state.

For our data, UTM zone 32N is the most sensible, but we set the
longitude offset for any UTM grid.

## Code changes:

- Adds branches to regrid._get_metadata and supporting functions with
tests