OG-46: Fetch grid metadata via STAC API and avoid redundant downloads (#72)
This PR introduces the following changes:

1. The static horizontal and vertical grid parameter files are now
retrieved via the STAC API using pre-signed URLs. Previously, these URLs
were hardcoded as a temporary solution due to missing metadata in the
STAC catalog. This logic is now integrated into both `get_from_ogd()`
and `download_from_ogd()`.
2. The `download_from_ogd()` function has been updated to automatically
download the corresponding horizontal and vertical grid parameter files
in addition to the requested forecast asset. This is necessary because
forecast GRIB files do not include height, latitude, or longitude
information, as the data is on an unstructured icosahedral grid.
3. A check was added to avoid re-downloading files that already exist in
the target folder, both for the main asset and the static grid files.


https://meteoswiss.atlassian.net/browse/OG-46?atlOrigin=eyJpIjoiZjE0NjA1YWQ0NDNjNDE5NTgyZDVkNzc1NzI0YmRhZGMiLCJwIjoiaiJ9