persist sha256 on disk and check (#78)
## Purpose

Persist the file sha256 hash on disk on a file with the same name and
the `.sha256` suffix. Lifecycle management of the files is left to the
user similarly to the downloaded files. The implementation should now be
able to identify partial downloads and replace the affected files but
skip downloads for files that comply with the checksum.