pass feature of polytope as dictionary (#118)
## Purpose

Keeping the extra dataclasses for client validation of the feature
extraction is always behind the capabilities of the server. Currently we
can not test some of the features that are already deployed and
available.
After discussion with the polytope team, we agreed that theses checks
can be implemented on the server (early, before the brokering)